requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
click>=8.0.0
pydantic>=1.10.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
flask>=2.0.0
flask-cors>=3.0.0
python-docx>=0.8.11
openai>=1.0.0
